separate the interfaces from component into own folder
