+++ App/interfaces/IArticleSource.ts
@@ -1,0 +1,6 @@
+interface IArticleSource {
+  id: string | null;
+  name: string;
+}
+
+export default IArticleSource;
