+++ App/interfaces/INews.ts
@@ -1,0 +1,9 @@
+import IArticles from 'interfaces/IArticles';
+
+interface INews {
+  status: string;
+  totalResults: number;
+  articles: Array<IArticles>;
+}
+
+export default INews;
